Workspace modal: tabbed Saved + New shell, inline Edit, Resume
First of three PRs implementing the Phase 2 (Actions) scope from
docs/design/workspace-modal.md (issue #56). Replaces the legacy
single-section CreateWorkspaceModal with a tabbed shell whose Saved
tab lets the user resume an existing workspace after editing it
inline; the New tab keeps the create flow on the same reusable form.

Out of scope for this PR (deferred to PR-B + PR-C):
- Clone + Delete row-footer actions
- Chip ⋮ menu Edit / Clone / Delete entries
- Restart-to-apply banner for live-workspace edits
- Advanced image search modal (magnifying-glass icon)

Components:
- WorkspaceForm.tsx — extracted, mode-aware form (create/edit). Owns
  field state, validation, footer. The submit shape carries both a
  full `secretKeys` list (lands in manifest.env.secretKeys) and a
  `secrets` map of newly-typed values (only these get setSecret'd).
  Pre-existing secret keys in edit mode render with a "•••••"
  placeholder; untouched ones flow through without rewriting keytar.
- WorkspaceModal.tsx — tabbed shell. Underlined-tab strip with Saved
  + New tabs. Default tab is Saved when any non-running workspace
  exists, else New. Saved tab includes Variant-B label-filter search:
  text input (matches name + description, case-insensitive) +
  Labels dropdown with usage counts + OR-filter pills with N-of-M
  count. Row click animates expansion via CSS grid-template-rows
  0fr→1fr 320ms cubic-bezier; chevron rotates 180°. Expanded body
  renders WorkspaceForm(mode='edit') with the spec pre-filled.

IPC:
- New `workspace:writeManifest(spec)` channel updates a manifest in
  place without touching the container. Resume calls it before
  workspace:start. Validates name-uniqueness (excluding own id).

App.tsx:
- handleRestart deleted; replaced by handleResume which writes secrets
  + manifest + start, then falls through to create-with-reused-id if
  the container is gone.

Tests:
- New tests/workspace-modal.spec.ts (6 specs) covers tab defaults,
  text search, labels OR filter + pills, row expansion, Resume
  triggering writeManifest + start in order.
- Existing create-flow + image-library tests swap the legacy
  `heading 'New workspace'` assertion for the New tab's role.
- workspace-lifecycle past-workspaces tests rewritten to use the
  Saved tab structure and assert Resume calls writeManifest + start
  rather than the old single-click restart.
- _helpers.ts stubs the new writeManifest channel.

Spec:
- docs/SPEC.md §5 (modal architecture), §6 (new IPC channel), §8
  (Saved tab default + Resume flow + Create-flow renumbering),
  §9 (env-row source), §10 (source tree). The Resume flow's
  3-case behaviour (stopped → start; deleted → recreate with id;
  container disappeared mid-flight → recreate) is documented in
  the user-flow section.

Refs #56 (partial — PR 1 of 3).
