+++ src/renderer/src/styles.css
@@ -1361,3 +1361,271 @@ body {
   font-family: var(--mono, ui-monospace, SFMono-Regular, Menlo, monospace);
   font-size: 13px;
 }
+
+/* ── WorkspaceModal: tabs + Saved tab list + search (Phase 2 PR-A) ────── */
+
+.modal-tabbed {
+  display: flex;
+  flex-direction: column;
+  width: min(720px, 92vw);
+  max-height: 88vh;
+  padding: 0;
+  overflow: hidden;
+}
+.modal-tabs {
+  display: flex;
+  gap: 0;
+  padding: 0 24px;
+  border-bottom: 1px solid var(--border, #2a2f38);
+  background: var(--bg-card, #181c24);
+}
+.modal-tab {
+  background: transparent;
+  border: none;
+  border-bottom: 2px solid transparent;
+  color: var(--text-muted);
+  font-size: 14px;
+  font-weight: 500;
+  padding: 12px 18px;
+  margin-bottom: -1px;
+  cursor: pointer;
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+}
+.modal-tab:hover { color: var(--text-primary); }
+.modal-tab.active {
+  color: var(--text-primary);
+  border-bottom-color: var(--accent, #5fa1de);
+}
+.modal-tab-count {
+  font-size: 11px;
+  padding: 1px 6px;
+  background: var(--bg-input, #11151c);
+  border-radius: 999px;
+  color: var(--text-muted);
+  font-weight: 400;
+}
+.modal-tab.active .modal-tab-count {
+  background: color-mix(in oklab, var(--accent, #5fa1de) 25%, transparent);
+  color: var(--text-primary);
+}
+
+.saved-tab,
+.new-tab {
+  padding: 18px 24px 20px;
+  overflow-y: auto;
+  flex: 1;
+}
+
+.saved-empty {
+  text-align: center;
+  padding: 32px 16px;
+  color: var(--text-muted);
+}
+.saved-empty p { margin: 0 0 12px; }
+
+.saved-search {
+  display: flex;
+  gap: 8px;
+  align-items: stretch;
+  margin-bottom: 8px;
+  position: relative;
+}
+.saved-search input[type="text"] {
+  flex: 1;
+  padding: 6px 10px;
+}
+.labels-filter { position: relative; }
+.labels-button {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  padding: 6px 12px;
+}
+.labels-button.active {
+  border-color: var(--accent, #5fa1de);
+  color: var(--text-primary);
+}
+.labels-count {
+  background: var(--accent, #5fa1de);
+  color: var(--bg, #0f1218);
+  font-size: 11px;
+  font-weight: 600;
+  padding: 0 6px;
+  border-radius: 999px;
+  min-width: 16px;
+  text-align: center;
+}
+.labels-dropdown {
+  position: absolute;
+  top: calc(100% + 4px);
+  right: 0;
+  z-index: 20;
+  min-width: 200px;
+  max-height: 280px;
+  overflow-y: auto;
+  background: var(--bg-card, #181c24);
+  border: 1px solid var(--border, #2a2f38);
+  border-radius: 8px;
+  padding: 6px;
+  box-shadow: 0 6px 24px rgba(0, 0, 0, 0.35);
+}
+.labels-dropdown-row {
+  display: grid;
+  grid-template-columns: auto 1fr auto;
+  align-items: center;
+  gap: 8px;
+  padding: 6px 8px;
+  font-size: 13px;
+  cursor: pointer;
+  border-radius: 4px;
+}
+.labels-dropdown-row:hover { background: var(--bg-hover, #1a1e26); }
+.labels-dropdown-count {
+  color: var(--text-muted);
+  font-size: 11px;
+}
+.labels-dropdown-footer {
+  border-top: 1px solid var(--border, #2a2f38);
+  margin-top: 6px;
+  padding-top: 6px;
+  text-align: right;
+}
+.labels-dropdown-clear {
+  background: none;
+  border: none;
+  color: var(--text-muted);
+  font-size: 12px;
+  cursor: pointer;
+  padding: 4px 6px;
+}
+.labels-dropdown-clear:hover { color: var(--text-primary); }
+
+.saved-active-filters {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 6px;
+  margin-bottom: 12px;
+  font-size: 12px;
+}
+.filter-pill {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  padding: 2px 6px 2px 10px;
+  background: color-mix(in oklab, var(--accent, #5fa1de) 18%, transparent);
+  border: 1px solid color-mix(in oklab, var(--accent, #5fa1de) 30%, transparent);
+  color: var(--text-primary);
+  border-radius: 999px;
+}
+.filter-pill button {
+  background: none;
+  border: none;
+  color: var(--text-muted);
+  cursor: pointer;
+  font-size: 14px;
+  padding: 0 2px;
+  line-height: 1;
+}
+.filter-pill button:hover { color: var(--state-error, #ef4d4d); }
+.saved-count {
+  margin-left: auto;
+  color: var(--text-muted);
+  font-size: 11px;
+}
+
+.saved-list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+.saved-row {
+  border: 1px solid var(--border, #2a2f38);
+  border-radius: 8px;
+  background: var(--bg-card, #181c24);
+  overflow: hidden;
+}
+.saved-row.expanded { border-color: var(--accent, #5fa1de); }
+.saved-row-header {
+  display: grid;
+  grid-template-columns: 6px 1fr auto auto;
+  align-items: center;
+  gap: 12px;
+  width: 100%;
+  padding: 10px 14px;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  color: var(--text-primary);
+  text-align: left;
+}
+.saved-row-header:hover { background: var(--bg-hover, #1a1e26); }
+.saved-row-color {
+  display: block;
+  width: 6px;
+  height: 28px;
+  border-radius: 3px;
+  align-self: center;
+}
+.saved-row-text {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  min-width: 0;
+}
+.saved-row-name {
+  font-weight: 500;
+  font-size: 14px;
+}
+.saved-row-desc {
+  color: var(--text-muted);
+  font-size: 12px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+.saved-row-meta {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  font-size: 11px;
+  color: var(--text-muted);
+}
+.saved-row-meta .ws-state {
+  text-transform: lowercase;
+  padding: 1px 6px;
+  border-radius: 4px;
+  font-size: 10px;
+  background: var(--bg-input, #11151c);
+}
+.saved-row-chevron {
+  color: var(--text-muted);
+  transition: transform 320ms cubic-bezier(0.32, 0.72, 0, 1);
+  font-size: 12px;
+}
+.saved-row-chevron.open { transform: rotate(-180deg); }
+
+.saved-row-body {
+  display: grid;
+  grid-template-rows: 0fr;
+  transition: grid-template-rows 320ms cubic-bezier(0.32, 0.72, 0, 1);
+}
+.saved-row-body.open { grid-template-rows: 1fr; }
+.saved-row-body-inner {
+  min-height: 0;
+  overflow: hidden;
+  opacity: 0;
+  transform: translateY(-3px);
+  transition: opacity 220ms 80ms ease, transform 220ms 80ms ease;
+  padding: 0 14px;
+}
+.saved-row-body.open .saved-row-body-inner {
+  opacity: 1;
+  transform: translateY(0);
+  padding: 4px 14px 14px;
+}
